Move MakerMacro to macros package
