+++ modules/akkaHttp/src/main/scala/ru/tinkoff/tschema/akkaHttp/MkRoute.scala
@@ -4,6 +4,7 @@ import akka.http.scaladsl.server.{Directive0, Directive1, Route}
 import ru.tinkoff.tschema.typeDSL.DSLDef
 import shapeless.HList
 import akka.http.scaladsl.server.Directives._
+import ru.tinkoff.tschema.macros.MakerMacro
 
 import language.experimental.macros
