+++ modules/macros/src/main/scala/ru/tinkoff/tschema/macros/MakerMacro.scala
@@ -1,4 +1,4 @@
-package ru.tinkoff.tschema.akkaHttp
+package ru.tinkoff.tschema.macros
 
 import java.time.Instant
